fix: append to build log rather than overwriting
Previously, tee was overwriting the logs for every invocation inside
build.sh. Use a parallel exec so that all output from the script goes to
the log. This means we no longer need a tee on each command, and we are
less likely to miss it off future additions. The build log may be very
long now, so we tail when grepping for the wheel names. The wheel name
should be in the last few lines of output at that point.

Also, change the date format to be more filname friendly (no spaces and
no colons)